requirements.txt
fastapi>=0.124.4
uvicorn>=0.38.0
pypdf2>=3.0.1
python-jose>=3.5.0
python-multipart>=0.0.20
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0
openai>=1.0.0
python-dateutil>=2.8.2
requests>=2.31.0
psycopg2-binary>=2.9.0
pinecone[grpc]>=4.0.0
reportlab>=4.0.0
slowapi>=0.1.8
celery>=5.3.0
redis>=5.0.0
tiktoken>=0.5.0
argon2-cffi>=23.0.0
